+++ src/iotjs_string.h
@@ -18,14 +18,14 @@
 
 
 typedef struct {
-  uint32_t size;
+  size_t size;
   char* data;
 } IOTJS_VALIDATED_STRUCT(iotjs_string_t);
 
 // Create new string
 iotjs_string_t iotjs_string_create();
-iotjs_string_t iotjs_string_create_with_size(const char* data, uint32_t size);
-iotjs_string_t iotjs_string_create_with_buffer(char* buffer, uint32_t size);
+iotjs_string_t iotjs_string_create_with_size(const char* data, size_t size);
+iotjs_string_t iotjs_string_create_with_buffer(char* buffer, size_t size);
 
 // Destroy string
 void iotjs_string_destroy(iotjs_string_t* str);
@@ -37,7 +37,7 @@ bool iotjs_string_is_empty(const iotjs_string_t* str);
 void iotjs_string_make_empty(iotjs_string_t* str);
 
 // Append `data` to tail of the string.
-void iotjs_string_append(iotjs_string_t* str, const char* data, uint32_t size);
+void iotjs_string_append(iotjs_string_t* str, const char* data, size_t size);
 
 // Returns pointer to the bytes (never returns NULL)
 const char* iotjs_string_data(const iotjs_string_t* str);
